backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai>=1.0.0
requests
python-dotenv
